New API is mostly done

+++ examples/make_fth_fig.py
@@ -53,7 +53,7 @@ if __name__ == '__main__':
 			Fmin,Fmax = 0.05,31
 			while abs(Fmax-Fmin) > tolerance:
 				F = np.exp((np.log(Fmax)+np.log(Fmin))/2)
-				t = np.linspace(-2*tau,2*tau,10*tau+1000)
+				t = np.linspace(-2*tau,2*tau,int(10*tau+1000))
 				laser = pp.Laser(wavelength=800e-9, pulse_duration=tau,fluence=F*1e4,t0=t.min(),transmit=True)
 				material = pp.Material(rate_equation=materials[m]["model"],index=materials[m]["index"],bandgap=materials[m]["bandgap"], \
 					m_CB=materials[m]["m_CB"], m_VB=materials[m]["m_VB"], density=materials[m]["density"], \
